Fix bug in quickstart
The download command in this quickstart fails because it attempts to download a release from the archived version of inlets, so inlets ends up just being an HTML 404 page. I replaced the commands for getting inlets with the commands from the SSH quickstart, as those commands did work for me.